Add coach_id column to coaches table under coaches_schema.sql and changes to the respective files like coach Entity, DTO, Repository, Service and Controller and coach_career Entity, DTO, Repository, Service and Controller

+++ src/main/kotlin/com/debrajghosh/nextxi/service/CoachCareerService.kt
@@ -15,7 +15,7 @@ class CoachCareerService(private val repository: CoachCareerRepository) {
         return repository.findById(id).orElse(null)?.let(CoachCareerDTO::fromObject)
     }
 
-    fun getCareersByCoach(coachId: Long): List<CoachCareerDTO> {
+    fun getCareersByCoach(coachId: Int): List<CoachCareerDTO> {
         return repository.findByCoachId(coachId).map(CoachCareerDTO::fromObject)
     }
